prototype/Python/pyproject.toml
[project]
name = "godo-lidar"
version = "0.1.0"
description = "GODO Phase 1 Python prototype for RPLIDAR C1 capture and noise characterization."
requires-python = ">=3.12"
dependencies = [
    "pyrplidar",
    "pyserial",
    "numpy",
    "pandas",
    "matplotlib",
]

[dependency-groups]
dev = [
    "pytest",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/godo_lidar"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra"
